使用GetInt替代部分GetFloat

+++ COG/Role/Impl/Crewmate/Vigilante.cs
@@ -39,7 +39,7 @@ public class Vigilante : CustomRole, IListener
     [OnlyLocalPlayerWithThisRoleInvokable]
     public void OnPlayerFixedUpdate(PlayerFixedUpdateEvent @event)
     {
-        if (PlayerUtils.AllCrewmates.Count() > _minCrewmateNumber.GetFloat() || _hasGiven) return;
+        if (PlayerUtils.AllCrewmates.Count() > _minCrewmateNumber.GetInt() || _hasGiven) return;
 
         DefaultKillButtonSetting.RemainingUses++;
         _hasGiven = true;
